pyproject.toml
[tool.poetry]
name = "project-3-group-ga-atm"
version = "0.1.0"
description = ""
authors = ["pnguyen121 <phillynguyen17@gmail.com>"]
readme = "README.md"
packages = [{include = "project_3_group_ga_atm"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
dj-database-url = "^1.2.0"
daphne = "^4.0.0"
psycopg2-binary = "^2.9.5"
boto3 = "^1.26.55"
geocoder = "^1.38.1"
django-crispy-forms = "^1.14.0"
environ = "^1.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
